% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling_functions.R
\name{wrangle_telonics}
\alias{wrangle_telonics}
\title{Sort through a directory of Telonics reports and read in a compiled dataset
of most recent broadcasts.}
\usage{
wrangle_telonics(path, tidy = TRUE)
}
\arguments{
\item{path}{Path of Telonics report direcory.}

\item{tidy}{Creates a tibble by default, \code{tidy = FALSE} returns a \code{data.frame}.}
}
\description{
Sort through a directory of Telonics reports and read in a compiled dataset
of most recent broadcasts.
}
\examples{
wrangle_telonics(path = "~/telonics/reports")
wrangle_telonics(path = "~/telonics/reports", tidy = FALSE)
}
